Upgrade to Laravel 5.1.1, upgrade other libraries
