--- src/app/Commands/Command.php
@@ -1,8 +1,0 @@
-<?php
-namespace TmlpStats\Commands;
-
-abstract class Command {
-
-    //
-
-}
