+++ src/app/Http/Controllers/Auth/AuthController.php
@@ -1,13 +1,14 @@
 <?php
 namespace TmlpStats\Http\Controllers\Auth;
 
+use TmlpStats\Person;
+use TmlpStats\User;
+use Validator;
 use TmlpStats\Http\Controllers\Controller;
-use Illuminate\Contracts\Auth\Guard;
-use Illuminate\Contracts\Auth\Registrar;
 use Illuminate\Foundation\Auth\AuthenticatesAndRegistersUsers;
 
-class AuthController extends Controller {
-
+class AuthController extends Controller
+{
     /*
     |--------------------------------------------------------------------------
     | Registration & Login Controller
@@ -24,16 +25,48 @@ class AuthController extends Controller {
     /**
      * Create a new authentication controller instance.
      *
-     * @param  \Illuminate\Contracts\Auth\Guard  $auth
-     * @param  \Illuminate\Contracts\Auth\Registrar  $registrar
-     * @return void
      */
-    public function __construct(Guard $auth, Registrar $registrar)
+    public function __construct()
     {
-        $this->auth = $auth;
-        $this->registrar = $registrar;
-
         $this->middleware('guest', ['except' => 'getLogout']);
     }
 
+    /**
+     * Get a validator for an incoming registration request.
+     *
+     * @param  array $data
+     * @return \Illuminate\Contracts\Validation\Validator
+     */
+    public function validator(array $data)
+    {
+        return Validator::make($data, [
+            'first_name'  => 'required|max:255',
+            'last_name'   => 'required|max:255',
+            'phone'       => 'regex:/^[\s\d\+\-\.]+$/',
+            'email'       => 'required|email|max:255|unique:users',
+            'password'    => 'required|confirmed|min:6',
+            'invite_code' => 'required|in:GloabalStatisticiansRock2015,GlobalStatisticiansRock2015',
+        ]);
+    }
+
+    /**
+     * Create a new user instance after a valid registration.
+     *
+     * @param  array $data
+     * @return \TmlpStats\User
+     */
+    public function create(array $data)
+    {
+        $person = Person::firstOrCreate([
+            'first_name' => $data['first_name'],
+            'last_name'  => $data['last_name'],
+            'email'      => $data['email'],
+        ]);
+
+        return User::create([
+            'person_id' => $person->id,
+            'email'     => $data['email'],
+            'password'  => bcrypt($data['password']),
+        ]);
+    }
 }
