+++ src/app/Http/Controllers/Auth/PasswordController.php
@@ -6,8 +6,8 @@ use Illuminate\Contracts\Auth\Guard;
 use Illuminate\Contracts\Auth\PasswordBroker;
 use Illuminate\Foundation\Auth\ResetsPasswords;
 
-class PasswordController extends Controller {
-
+class PasswordController extends Controller
+{
     /*
     |--------------------------------------------------------------------------
     | Password Reset Controller
@@ -24,16 +24,9 @@ class PasswordController extends Controller {
     /**
      * Create a new password controller instance.
      *
-     * @param  \Illuminate\Contracts\Auth\Guard  $auth
-     * @param  \Illuminate\Contracts\Auth\PasswordBroker  $passwords
-     * @return void
      */
-    public function __construct(Guard $auth, PasswordBroker $passwords)
+    public function __construct()
     {
-        $this->auth = $auth;
-        $this->passwords = $passwords;
-
         $this->middleware('guest');
     }
-
 }
