+++ src/composer.json
@@ -5,10 +5,10 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "laravel/framework": "5.0.*",
-        "laravelcollective/html": "~5.0",
-        "kris/laravel-form-builder": "~1.3",
-        "kirkbushell/eloquence": "~1.2",
+        "laravel/framework": "5.1.*",
+        "laravelcollective/html": "~5.1",
+        "kris/laravel-form-builder": "~1.6",
+        "kirkbushell/eloquence": "~1.3",
         "nesbot/carbon": "~1.14",
         "respect/validation": "~0.6",
         "phpoffice/phpexcel": "~1.8",
